requirements.txt
uvicorn==0.23.0
fastapi==0.100.0
tortoise-orm==0.19.3
asyncpg==0.28.0
requests==2.31.0
python-multipart==0.0.6